refactor: fix ImageUploader and PieceNameEditor prop boundaries
ImageUploader now owns the full Cloudinary upload lifecycle (absorbs
useImageSaveQueue). Callers pass configuration (pieceId, saveStateFn,
images, etc.) instead of transient loading/error state. useImageSaveQueue.ts
is deleted — its logic lives inside ImageUploader.

PieceNameEditor drops the redundant canEdit prop and derives it from
piece.can_edit directly. The parent already has piece; there is no reason
to re-derive and re-forward a boolean that the component can compute itself.

Test: updated WorkflowState test to query hidden button by role (not the
removed HTML hidden attribute) and simplified the read-only guard test.
