+++ web/src/components/__tests__/WorkflowState.test.tsx
@@ -841,14 +841,10 @@ describe("WorkflowState", () => {
     expect(api.fetchCloudinaryWidgetConfig).not.toHaveBeenCalled();
   });
 
-  it("returns before opening the upload widget when read-only", async () => {
+  it("does not open the upload widget when read-only", async () => {
     await act(async () => {
       render(<WorkflowState {...defaultProps} readOnly />);
     });
-
-    const hiddenUploadButton = document.querySelector("button[hidden]");
-    expect(hiddenUploadButton).toBeInstanceOf(HTMLButtonElement);
-    fireEvent.click(hiddenUploadButton as HTMLButtonElement);
 
     expect(api.fetchCloudinaryWidgetConfig).not.toHaveBeenCalled();
   });
